chore: Refactor web functionality out of client
